+++ docs/whatsnew.md
@@ -2,6 +2,7 @@
 
 *V1.2.0*
 - add Object.evmask to change link descriptor event mask
+- Variable.valueChanged now accepts a callback with signature ( value, Object )
 - fix: Variable/Task/Module c'tor: check if 'name' is equal to name in 'CD'
 - fix: Variable/Task/Module c'tor: content of 'CD' was overwritten with 'name'